requirement.txt
asgiref==3.2.10
boto3==1.14.7
botocore==1.17.7
certifi==2020.6.20
chardet==3.0.4
Django==3.0.7
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
docutils==0.15.2
idna==2.9
jmespath==0.10.0
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
